+++ Product/Production/Services/PatientDiscoveryCore/src/main/java/gov/hhs/fha/nhinc/patientdb/dao/PhonenumberDAO.java
@@ -1,61 +1,60 @@
 /*
- * Copyright (c) 2012, United States Government, as represented by the Secretary of Health and Human Services. 
- * All rights reserved. 
+ * Copyright (c) 2009-2016, United States Government, as represented by the Secretary of Health and Human Services.
+ * All rights reserved.
  *
- * Redistribution and use in source and binary forms, with or without 
- * modification, are permitted provided that the following conditions are met: 
- *     * Redistributions of source code must retain the above 
- *       copyright notice, this list of conditions and the following disclaimer. 
- *     * Redistributions in binary form must reproduce the above copyright 
- *       notice, this list of conditions and the following disclaimer in the documentation 
- *       and/or other materials provided with the distribution. 
- *     * Neither the name of the United States Government nor the 
- *       names of its contributors may be used to endorse or promote products 
- *       derived from this software without specific prior written permission. 
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *     * Redistributions of source code must retain the above
+ *       copyright notice, this list of conditions and the following disclaimer.
+ *     * Redistributions in binary form must reproduce the above copyright
+ *       notice, this list of conditions and the following disclaimer in the documentation
+ *       and/or other materials provided with the distribution.
+ *     * Neither the name of the United States Government nor the
+ *       names of its contributors may be used to endorse or promote products
+ *       derived from this software without specific prior written permission.
  *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND 
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED 
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE 
- * DISCLAIMED. IN NO EVENT SHALL THE UNITED STATES GOVERNMENT BE LIABLE FOR ANY 
- * DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES 
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; 
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND 
- * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT 
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS 
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE. 
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE UNITED STATES GOVERNMENT BE LIABLE FOR ANY
+ * DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 package gov.hhs.fha.nhinc.patientdb.dao;
 
 import gov.hhs.fha.nhinc.patientdb.model.Phonenumber;
 import gov.hhs.fha.nhinc.patientdb.persistence.HibernateUtil;
-
+import gov.hhs.fha.nhinc.patientdb.persistence.HibernateUtilFactory;
 import java.util.List;
-
-import org.apache.log4j.Logger;
 import org.hibernate.Criteria;
+import org.hibernate.HibernateException;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 import org.hibernate.Transaction;
 import org.hibernate.criterion.Expression;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
- * 
+ *
  * PhonenumberDAO Class provides methods to query and update Phonenumber Data to/from MySQL Database using Hibernate
- * 
+ *
  * @author richard.ettema
  */
-
 public class PhonenumberDAO {
 
-    private static final Logger LOG = Logger.getLogger(PhonenumberDAO.class);
+    private static final Logger LOG = LoggerFactory.getLogger(PhonenumberDAO.class);
 
     private static PhonenumberDAO phonenumberDAO = new PhonenumberDAO();
 
     /**
-     * 
+     *
      * Constructor
      */
-
     private PhonenumberDAO() {
 
         LOG.info("PhonenumberDAO - Initialized");
@@ -63,12 +62,11 @@ public class PhonenumberDAO {
     }
 
     /**
-     * 
+     *
      * Singleton instance returned...
-     * 
+     *
      * @return PhonenumberDAO
      */
-
     public static PhonenumberDAO getPhonenumberDAOInstance() {
 
         LOG.debug("getPhonenumberDAOInstance()..");
@@ -78,37 +76,33 @@ public class PhonenumberDAO {
     }
 
     // =========================
-
     // Standard CRUD DML Methods
-
     // =========================
-
     /**
-     * 
+     *
      * Create a single <code>Phonenumber</code> record. The generated id
-     * 
+     *
      * will be available in the phonenumberRecord.
-     * 
+     *
      * @param phonenumberRecord
-     * 
+     *
      * @return boolean
      */
-
     public boolean create(Phonenumber phonenumberRecord) {
 
         LOG.debug("PhonenumberDAO.create() - Begin");
 
         Session session = null;
 
-        Transaction tx = null;
-
         boolean result = true;
 
         if (phonenumberRecord != null) {
 
+            Transaction tx = null;
+
             try {
 
-                SessionFactory sessionFactory = HibernateUtil.getSessionFactory();
+                SessionFactory sessionFactory = getSessionFactory();
 
                 session = sessionFactory.openSession();
 
@@ -122,7 +116,7 @@ public class PhonenumberDAO {
 
                 tx.commit();
 
-            } catch (Exception e) {
+            } catch (HibernateException | NullPointerException e) {
 
                 result = false;
 
@@ -132,18 +126,18 @@ public class PhonenumberDAO {
 
                 }
 
-                LOG.error("Exception during insertion caused by :" + e.getMessage(), e);
+                LOG.error("Exception during insertion caused by : {}", e.getMessage(), e);
 
             } finally {
 
                 // Actual Phonenumber insertion will happen at this step
-
                 if (session != null) {
-
-                    session.close();
-
+                    try {
+                        session.close();
+                    } catch (HibernateException e) {
+                        LOG.error("Exception while closing the session: {}", e.getMessage(), e);
+                    }
                 }
-
             }
 
         }
@@ -155,16 +149,15 @@ public class PhonenumberDAO {
     }
 
     /**
-     * 
+     *
      * Read (Query) the database to get a <code>Phonenumber</code> record based
-     * 
+     *
      * on a known id.
-     * 
+     *
      * @param id
-     * 
+     *
      * @return Phonenumber
      */
-
     public Phonenumber read(Long id) {
 
         LOG.debug("PhonenumberDAO.read() - Begin");
@@ -181,48 +174,49 @@ public class PhonenumberDAO {
 
         Session session = null;
 
-        List<Phonenumber> queryList = null;
+        List<Phonenumber> queryList;
 
         Phonenumber foundRecord = null;
 
         try {
 
-            SessionFactory sessionFactory = HibernateUtil.getSessionFactory();
+            SessionFactory sessionFactory = getSessionFactory();
 
-            session = sessionFactory.openSession();
+            if (sessionFactory != null) {
+                session = sessionFactory.openSession();
 
-            LOG.info("Reading Record...");
+                LOG.info("Reading Record...");
 
-            // Build the criteria
+                // Build the criteria
+                Criteria aCriteria = session.createCriteria(Phonenumber.class);
 
-            Criteria aCriteria = session.createCriteria(Phonenumber.class);
+                aCriteria.add(Expression.eq("id", id));
 
-            aCriteria.add(Expression.eq("id", id));
+                queryList = aCriteria.list();
 
-            queryList = aCriteria.list();
+                if (queryList != null && !queryList.isEmpty()) {
 
-            if (queryList != null && queryList.size() > 0) {
+                    foundRecord = queryList.get(0);
 
-                foundRecord = queryList.get(0);
-
+                }
             }
 
-        } catch (Exception e) {
+        } catch (HibernateException | NullPointerException e) {
 
-            LOG.error("Exception during read occured due to :" + e.getMessage(), e);
+            LOG.error("Exception during read occured due to : {}", e.getMessage(), e);
 
         } finally {
 
             // Flush and close session
-
             if (session != null) {
 
-                session.flush();
-
-                session.close();
-
+                try {
+                    session.flush();
+                    session.close();
+                } catch (HibernateException e) {
+                    LOG.error("Exception while closing the session after a read: {}", e.getMessage(), e);
+                }
             }
-
         }
 
         LOG.debug("PhonenumberDAO.read() - End");
@@ -232,29 +226,28 @@ public class PhonenumberDAO {
     }
 
     /**
-     * 
+     *
      * Update a single <code>Phonenumber</code> record.
-     * 
+     *
      * @param phonenumberRecord
-     * 
+     *
      * @return boolean
      */
-
     public boolean update(Phonenumber phonenumberRecord) {
 
         LOG.debug("PhonenumberDAO.update() - Begin");
 
         Session session = null;
 
-        Transaction tx = null;
-
         boolean result = true;
 
         if (phonenumberRecord != null) {
 
+            Transaction tx = null;
+
             try {
 
-                SessionFactory sessionFactory = HibernateUtil.getSessionFactory();
+                SessionFactory sessionFactory = getSessionFactory();
 
                 session = sessionFactory.openSession();
 
@@ -268,7 +261,7 @@ public class PhonenumberDAO {
 
                 tx.commit();
 
-            } catch (Exception e) {
+            } catch (HibernateException | NullPointerException e) {
 
                 result = false;
 
@@ -278,18 +271,18 @@ public class PhonenumberDAO {
 
                 }
 
-                LOG.error("Exception during update caused by :" + e.getMessage(), e);
+                LOG.error("Exception during update caused by : {}", e.getMessage(), e);
 
             } finally {
 
                 // Actual Phonenumber update will happen at this step
-
                 if (session != null) {
-
-                    session.close();
-
+                    try {
+                        session.close();
+                    } catch (HibernateException e) {
+                        LOG.error("Exception while closing the session after an update: {}", e.getMessage(), e);
+                    }
                 }
-
             }
 
         }
@@ -301,12 +294,11 @@ public class PhonenumberDAO {
     }
 
     /**
-     * 
+     *
      * Delete a <code>Phonenumber</code> record from the database
-     * 
+     *
      * @param phonenumberRecord
      */
-
     public void delete(Phonenumber phonenumberRecord) {
 
         LOG.debug("PhonenumberDAO.delete() - Begin");
@@ -315,32 +307,33 @@ public class PhonenumberDAO {
 
         try {
 
-            SessionFactory sessionFactory = HibernateUtil.getSessionFactory();
+            SessionFactory sessionFactory = getSessionFactory();
 
-            session = sessionFactory.openSession();
+            if (sessionFactory != null) {
+                session = sessionFactory.openSession();
 
-            LOG.info("Deleting Record...");
+                LOG.info("Deleting Record...");
 
-            // Delete the Phonenumber record
+                // Delete the Phonenumber record
+                session.delete(phonenumberRecord);
+            }
 
-            session.delete(phonenumberRecord);
+        } catch (HibernateException | NullPointerException e) {
 
-        } catch (Exception e) {
-
-            LOG.error("Exception during delete occured due to :" + e.getMessage(), e);
+            LOG.error("Exception during delete occured due to : {}", e.getMessage(), e);
 
         } finally {
 
             // Flush and close session
-
             if (session != null) {
 
-                session.flush();
-
-                session.close();
-
+                try {
+                    session.flush();
+                    session.close();
+                } catch (HibernateException e) {
+                    LOG.error("Exception while closing the session after a delete: {}", e.getMessage(), e);
+                }
             }
-
         }
 
         LOG.debug("PhonenumberDAO.delete() - End");
@@ -348,22 +341,18 @@ public class PhonenumberDAO {
     }
 
     // =========================
-
     // Find DML Methods
-
     // =========================
-
     /**
-     * 
+     *
      * Read (Query) the database to get a <code>Phonenumber</code> record based
-     * 
+     *
      * on a known patientId.
-     * 
+     *
      * @param patientId
-     * 
+     *
      * @return List<Phonenumber>
      */
-
     public List<Phonenumber> findPatientPhonenumbers(Long patientId) {
 
         LOG.debug("PhonenumberDAO.findPatientPhonenumbers() - Begin");
@@ -384,42 +373,57 @@ public class PhonenumberDAO {
 
         try {
 
-            SessionFactory sessionFactory = HibernateUtil.getSessionFactory();
+            SessionFactory sessionFactory = getSessionFactory();
 
-            session = sessionFactory.openSession();
+            if (sessionFactory != null) {
+                session = sessionFactory.openSession();
 
-            LOG.info("Reading Record...");
+                LOG.info("Reading Record...");
 
-            // Build the criteria
+                // Build the criteria
+                Criteria aCriteria = session.createCriteria(Phonenumber.class);
 
-            Criteria aCriteria = session.createCriteria(Phonenumber.class);
+                aCriteria.add(Expression.eq("patient.patientId", patientId));
 
-            aCriteria.add(Expression.eq("patient.patientId", patientId));
+                queryList = aCriteria.list();
 
-            queryList = aCriteria.list();
+            }
+        } catch (HibernateException | NullPointerException e) {
 
-        } catch (Exception e) {
-
-            LOG.error("Exception during read occured due to :" + e.getMessage(), e);
+            LOG.error("Exception during read occured due to : {}", e.getMessage(), e);
 
         } finally {
 
             // Flush and close session
-
             if (session != null) {
-
-                session.flush();
-
-                session.close();
-
+                try {
+                    session.flush();
+                    session.close();
+                } catch (HibernateException e) {
+                    LOG.error("Exception while closing the session after looking for patient's phone numbers: {}",
+                            e.getMessage(), e);
+                }
             }
-
         }
 
         LOG.debug("PhonenumberDAO.findPatientPhonenumbers() - End");
 
         return queryList;
 
+    }
+
+    /**
+     * Returns the sessionFactory belonging to PatientDiscovery HibernateUtil
+     *
+     * @return
+     */
+    protected SessionFactory getSessionFactory() {
+        SessionFactory fact = null;
+        HibernateUtil hibernateUtil = HibernateUtilFactory.getPatientDiscHibernateUtil();
+        if (hibernateUtil != null) {
+            fact = hibernateUtil.getSessionFactory();
+        }
+        return fact;
     }
 
 }
